requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Web framework and ASGI server
fastapi>=0.104.0
uvicorn>=0.24.0

# Dependency resolution
pipgrip>=0.10.0

# Async HTTP client
aiohttp>=3.8.0
aiofiles>=23.0.0

# Data validation and serialization
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.21.0

openpyxl

uv-dep-resolver